Added subscriptionService and made it globally available
